update: 로비 관련 함수 추가, 접두어 변경, 만료시간 속성화
기타 함수명 오타 수정, room 데이터 구조 변경,
lobbyId 추가

+++ packages/common/classes/models/redisUtil.class.js
@@ -2,29 +2,11 @@
  * @typedef userdata
  * @property {string} loginId
  * @property {string} nickname
- * @property {string} location
- */
-
-/**
- * @typedef userField
- * @property {string} loginId
- * @property {string} nickname
- * @property {string} location
- */
-
-/**
- * @typedef roomField
- * @property {string} roomId
- * @property {string} ownerId
- * @property {string} roomName
- * @property {string} state
- * @property {number} maxUser
- * @property {set} readyUsers
  */
 
 /**
  * @typedef locationField
- * @property {string} lobby
+ * @property {string} lobbyId
  * @property {string} roomId
  * @property {string} boardId
  * @property {string} miniGameId
@@ -35,9 +17,11 @@
  * @property {string} roomId
  * @property {string} ownerSessionId
  * @property {string} roomName
+ * @property {string} lobbyId
  * @property {string} state
- * @property {number} maxUser
- * @property {set} readyUsers
+ * @property {string} users
+ * @property {string} maxUser
+ * @property {string} readyUsers
  */
 
 class RedisUtil {
@@ -47,11 +31,13 @@ class RedisUtil {
       USER: 'user',
       LOCATION: 'location',
       LOBBY_USERS: 'lobbyUsers',
-      ROOM_LIST: 'roomList',
+      LOBBY_ROOM_LIST: 'lobbyRoomList',
       ROOM: 'room',
     };
+    console.log('Redis prefix:', this.prefix);
+    this.expire = 60 * 60;
   }
-  // 유저 위치
+  //* 유저 위치
   /**
    * 유저의 위치 데이터 등록
    * @param {string} sessionId
@@ -61,7 +47,7 @@ class RedisUtil {
   async createUserLocation(sessionId, location, locationId) {
     const key = `${this.prefix.LOCATION}:${sessionId}`;
     await this.client.hset(key, location, locationId);
-    await this.client.expire(key, 60 * 60);
+    await this.client.expire(key, this.expire);
   }
 
   /**
@@ -92,7 +78,7 @@ class RedisUtil {
   async updateUserLocationField(sessionId, locationField, locationId) {
     const key = `${this.prefix.LOCATION}:${sessionId}`;
     await this.client.hset(key, locationField, locationId);
-    await this.client.expire(key, 60 * 60);
+    await this.client.expire(key, this.expire);
   }
 
   /**
@@ -118,7 +104,7 @@ class RedisUtil {
    */
   async getUserLocation(sessionId) {
     const key = `${this.prefix.LOCATION}:${sessionId}`;
-    const data = await this.client.hget(key);
+    const data = await this.client.hgetall(key);
     if (!data) {
       return null;
     }
@@ -126,7 +112,7 @@ class RedisUtil {
     return data;
   }
 
-  // 유저 데이터
+  //* 유저 데이터
   /**
    * 세션에 유저 데이터 등록
    * @param {string} sessionId
@@ -137,9 +123,8 @@ class RedisUtil {
     await this.client.hset(key, {
       loginId: userData.loginId,
       nickname: userData.nickname,
-      location: userData.location,
     });
-    await this.client.expire(key, 60 * 60);
+    await this.client.expire(key, this.expire);
   }
 
   /**
@@ -154,19 +139,19 @@ class RedisUtil {
   /**
    * 유저 데이터 중 하나의 필드만 수정
    * @param {string} sessionId
-   * @param {userField} userField
+   * @param {userdata} userField
    * @param {string} value
    */
   async updateUserToSessionfield(sessionId, userField, value) {
     const key = `${this.prefix.USER}:${sessionId}`;
     await this.client.hset(key, userField, value);
-    await this.client.expire(key, 60 * 60);
+    await this.client.expire(key, this.expire);
   }
 
   /**
    * 유저 데이터 중 하나의 필드만 조회
    * @param {string} sessionId
-   * @param {userField} userField
+   * @param {userdata} userField
    * @returns {string} data
    */
   async getUserToSessionfield(sessionId, userField) {
@@ -186,7 +171,7 @@ class RedisUtil {
    */
   async getUserToSession(sessionId) {
     const key = `${this.prefix.USER}:${sessionId}`;
-    const userData = await this.client.hget(key);
+    const userData = await this.client.hgetall(key);
     if (!userData) {
       return null;
     }
@@ -194,26 +179,28 @@ class RedisUtil {
     return {
       loginId: userData.loginId,
       nickname: userData.nickname,
-      location: userData.location,
     };
   }
 
-  // 로비 데이터
+  //* 로비 데이터
   /**
    * 로비에 접속한 유저 추가
    * @param {string} sessionId
+   * @param {string} lobbyId
    */
-  async addLobbyUsers(sessionId) {
-    const key = `${this.prefix.LOBBY_USERS}`;
+  async addLobbyUsers(sessionId, lobbyId) {
+    const key = `${this.prefix.LOBBY_USERS}:${lobbyId}`;
     await this.client.sadd(key, sessionId);
+    await this.client.expire(key, this.expire);
   }
 
   /**
    * 로비에 접속 중인 유저 목록 조회
+   * @param {string} lobbyId - 로비 ID
    * @returns {Array<userdata>} userData
    */
-  async getLobbyUserss() {
-    const key = `${this.prefix.LOBBY_USERS}`;
+  async getLobbyUsers(lobbyId) {
+    const key = `${this.prefix.LOBBY_USERS}:${lobbyId}`;
     const users = await this.client.smembers(key);
     if (!users) {
       return null;
@@ -222,44 +209,110 @@ class RedisUtil {
     return users;
   }
 
-  // 룸 데이터
+  /**
+   * 로비에 접속 중인 유저 삭제
+   * @param {string} sessionId
+   * @param {string} lobbyId
+   */
+  async deleteLobbyUser(sessionId, lobbyId) {
+    const key = `${this.prefix.LOBBY_USERS}:${lobbyId}`;
+    await this.client.srem(key, sessionId);
+  }
+
+  //* 룸 데이터
+
+  async getRoomsByLobby(lobbyId) {
+    const roomKeys = await this.client.smembers(`${this.prefix.LOBBY_ROOM_LIST}:${lobbyId}`);
+    if (!roomKeys || roomKeys.length === 0) return [];
+
+    const roomIds = roomKeys.map((key) => key.replace(`${this.prefix.ROOM}:`, ''));
+    const rooms = await Promise.all(roomIds.map((roomId) => this.getRoom(roomId)));
+
+    return rooms.filter((room) => room !== null);
+  }
+
   /**
    * 대기방 생성
    * @param {room} room
    * @param {string} sessionId
    */
-  async createRoom(room, sessionId) {
+  async createRoom(room) {
     const key = `${this.prefix.ROOM}:${room.roomId}`;
+    const exists = await this.client.exists(key);
+    if (exists) {
+      return false;
+    }
+
     await this.client.hset(key, {
-      ownerId: sessionId,
+      roomId: room.roomId,
+      ownerId: room.ownerId,
       name: room.name,
       state: room.state,
+      lobbyId: room.lobbyId,
+      users: room.users,
       maxUser: room.maxUser,
       readyUsers: room.readyUsers,
     });
-    await this.client.expire(key, 60 * 60);
-    await this.client.sadd(`${this.prefix.ROOM_LIST}`, key);
+    await this.client.expire(key, this.expire);
+    await this.client.sadd(`${this.prefix.LOBBY_ROOM_LIST}:${room.lobbyId}`, key);
+  }
+
+  /**
+   * 대기방 정보 조회
+   * @param {string} roomId
+   * @returns {room} room
+   */
+  async getRoom(roomId) {
+    const key = `${this.prefix.ROOM}:${roomId}`;
+    console.log('Redis Key:', key);
+    const roomData = await this.client.hgetall(key);
+    if (!roomData) {
+      return null;
+    }
+
+    return {
+      roomId: roomData.roomId,
+      ownerId: roomData.ownerId,
+      name: roomData.name,
+      state: roomData.state,
+      lobbyId: roomData.lobbyId,
+      users: roomData.users,
+      maxUser: roomData.maxUser,
+      readyUsers: roomData.readyUsers,
+    };
   }
 
   /**
    * 대기방의 하나의 필드만 수정
    * @param {string} roomId
-   * @param {roomField} roomField
-   * @param {string | number | set} value
+   * @param {room} roomField
+   * @param {string} value
    */
-  async updateRoomfield(roomId, roomField, value) {
+  async updateRoomField(roomId, roomField, value) {
     const key = `${this.prefix.ROOM}:${roomId}`;
     await this.client.hset(key, roomField, value);
-    await this.client.expire(key, 60 * 60);
+    await this.client.expire(key, this.expire);
+  }
+
+  /**
+   *대기방의 여러개의 필드 수정
+   * @param {string} roomId
+   * @param {room} roomFields
+   * @returns
+   */
+  async updateRoomFields(roomId, roomFields) {
+    const key = `${this.prefix.ROOM}:${roomId}`;
+    await this.client.hset(key, roomFields);
+    await this.client.expire(key, this.expire);
   }
 
   /**
    * 대기방의 하나의 필드의 값 조회
    * @param {string} roomId
-   * @param {roomField} roomField
-   * @returns {string | number | set} data
+   * @param {room} roomField
+   * @returns {string} data
    */
-  async getRoomfield(roomId, roomField) {
+  async getRoomField(roomId, roomField) {
     const key = `${this.prefix.ROOM}:${roomId}`;
     const data = await this.client.hget(key, roomField);
     if (!data) {
@@ -275,7 +328,13 @@ class RedisUtil {
    */
   async deleteRoom(roomId) {
     const key = `${this.prefix.ROOM}:${roomId}`;
-    await this.client.del(key);
+    const roomData = await this.getRoom(roomId);
+    if (roomData) {
+      await Promise.all([
+        this.client.del(key),
+        this.client.srem(`${this.prefix.LOBBY_ROOM_LIST}:${roomData.lobbyId}`, key),
+      ]);
+    }
   }
 }
